+++ src/components/product-card/product-card.component.tsx
@@ -3,11 +3,11 @@ import './product-card.style.scss';
 import { useDispatch, useSelector } from 'react-redux';
 import { addItemToCart } from '../../store/cart/cart.action';
 import { selectCartItems } from '../../store/cart/cart.selector';
-import { CartItem } from '../../store/cart/cart.types';
 import { FC } from 'react';
+import { CategoryItem } from '../../store/categories/categories.types';
 
 type ProductCardProps = {
-  product: CartItem;
+  product: CategoryItem;
 };
 const ProductCard: FC<ProductCardProps> = ({ product }) => {
   const dispatch = useDispatch();
