+++ src/routes/shop/shop.component.tsx
